svg/shapes/line-pathLength.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><metadata><h:link href="https://www.w3.org/TR/SVG2/shapes.html#InterfaceSVGLineElement" rel="help"/><h:link href="https://www.w3.org/TR/SVG2/types.html#InterfaceSVGGeometryElement" rel="help"/><h:meta content="The pathLength IDL attribute reflects the ‘pathLength’ content attribute." name="assert"/></metadata><line id="interval" x1="300" x2="500" y1="400" y2="400" pathLength="25"/><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script><![CDATA[
  test(function() {
    var interval = document.getElementById('interval');
    assert_true(interval.pathLength !== undefined, "pathLength is defined for SVGLineElement");
    assert_equals(interval.pathLength.baseVal, 25);
  }, 'line supports pathLength attribute');
  ]]></script></svg>
